make build pass, switch to local payload fetch internally
