+++ src/components/content/formBlock.tsx
@@ -3,7 +3,7 @@ import React, { FC, Fragment, useState } from 'react'
 import { useForm } from 'react-hook-form'
 
 import { Form as FormTypes } from '@/payload-types'
-import { serverUrl } from '@/utilities/api'
+import { serverUrl } from '@/utilities/serverConfig'
 import { Block } from '@/components/ui/block'
 import { Button } from '@/components/ui/button'
 import { Dialog, DialogContent } from '@/components/ui/dialog'
@@ -151,7 +151,7 @@ export const FormBlock: FC<FormBlockProps> = props => {
 
               const props = {
                 id: `${formID}-${field.name}`,
-                ...register(field.name, { required: field.required, pattern }),
+                ...register(field.name, { required: Boolean(field.required), pattern }),
               }
 
               let content
