+++ src/payload-types.ts
@@ -15,368 +15,496 @@ export interface Config {
     users: User;
     forms: Form;
     'form-submissions': FormSubmission;
+    'payload-preferences': PayloadPreference;
+    'payload-migrations': PayloadMigration;
   };
   globals: {
     header: Header;
     profile: Profile;
   };
+  locale: null;
+  user: User & {
+    collection: 'users';
+  };
 }
+/**
+ * This interface was referenced by `Config`'s JSON-Schema
+ * via the `definition` "media".
+ */
 export interface Media {
   id: string;
   alt: string;
   updatedAt: string;
   createdAt: string;
-  url?: string;
-  filename?: string;
-  mimeType?: string;
-  filesize?: number;
-  width?: number;
-  height?: number;
+  url?: string | null;
+  thumbnailURL?: string | null;
+  filename?: string | null;
+  mimeType?: string | null;
+  filesize?: number | null;
+  width?: number | null;
+  height?: number | null;
+  focalX?: number | null;
+  focalY?: number | null;
 }
+/**
+ * This interface was referenced by `Config`'s JSON-Schema
+ * via the `definition` "pages".
+ */
 export interface Page {
   id: string;
   title: string;
-  layout?: (
-    | {
-        contentFields?: {
-          size?: 'oneThird' | 'half' | 'twoThirds' | 'full';
-          richText?: {
-            [k: string]: unknown;
-          }[];
-          enableLink?: boolean;
-          link?: {
-            type?: 'reference' | 'custom';
-            newTab?: boolean;
-            reference:
+  layout?:
+    | (
+        | {
+            contentFields?:
               | {
-                  value: string | Page;
-                  relationTo: 'pages';
-                }
+                  size?: ('oneThird' | 'half' | 'twoThirds' | 'full') | null;
+                  richText?:
+                    | {
+                        [k: string]: unknown;
+                      }[]
+                    | null;
+                  enableLink?: boolean | null;
+                  link?: {
+                    type?: ('reference' | 'custom') | null;
+                    newTab?: boolean | null;
+                    reference?:
+                      | ({
+                          relationTo: 'pages';
+                          value: string | Page;
+                        } | null)
+                      | ({
+                          relationTo: 'projects';
+                          value: string | Project;
+                        } | null);
+                    url?: string | null;
+                    label: string;
+                    appearance?: ('primary' | 'secondary') | null;
+                  };
+                  id?: string | null;
+                }[]
+              | null;
+            id?: string | null;
+            blockName?: string | null;
+            blockType: 'content';
+          }
+        | {
+            richText?:
               | {
-                  value: string | Project;
-                  relationTo: 'projects';
-                };
-            url: string;
-            label: string;
-            appearance?: 'primary' | 'secondary';
-          };
-          id?: string;
-        }[];
-        id?: string;
-        blockName?: string;
-        blockType: 'content';
-      }
-    | {
-        richText?: {
-          [k: string]: unknown;
-        }[];
-        form: string | Form;
-        id?: string;
-        blockName?: string;
-        blockType: 'form';
-      }
-    | {
-        mediaFields?: {
-          size?: 'oneThird' | 'half' | 'twoThirds' | 'full';
-          media: string | Media;
-          mediaFit?: 'cover' | 'contain';
-          id?: string;
-        }[];
-        id?: string;
-        blockName?: string;
-        blockType: 'mediaBlock';
-      }
-    | {
-        mediaContentFields?: {
-          alignment?: 'contentMedia' | 'mediaContent';
-          mediaSize?: 'oneThird' | 'half' | 'twoThirds';
-          richText?: {
-            [k: string]: unknown;
-          }[];
-          enableLink?: boolean;
-          link?: {
-            type?: 'reference' | 'custom';
-            newTab?: boolean;
-            reference:
+                  [k: string]: unknown;
+                }[]
+              | null;
+            form: string | Form;
+            id?: string | null;
+            blockName?: string | null;
+            blockType: 'form';
+          }
+        | {
+            mediaFields?:
               | {
-                  value: string | Page;
-                  relationTo: 'pages';
-                }
+                  size?: ('oneThird' | 'half' | 'twoThirds' | 'full') | null;
+                  media: string | Media;
+                  mediaFit?: ('cover' | 'contain') | null;
+                  id?: string | null;
+                }[]
+              | null;
+            id?: string | null;
+            blockName?: string | null;
+            blockType: 'mediaBlock';
+          }
+        | {
+            mediaContentFields?:
               | {
-                  value: string | Project;
-                  relationTo: 'projects';
-                };
-            url: string;
-            label: string;
-            appearance?: 'primary' | 'secondary';
-          };
-          media: string | Media;
-          mediaFit?: 'cover' | 'contain';
-          id?: string;
-        }[];
-        id?: string;
-        blockName?: string;
-        blockType: 'mediaContent';
-      }
-    | {
-        id?: string;
-        blockName?: string;
-        blockType: 'profile-cta';
-      }
-    | {
-        project: string[] | Project[];
-        id?: string;
-        blockName?: string;
-        blockType: 'projectGrid';
-      }
-  )[];
-  slug?: string;
+                  alignment?: ('contentMedia' | 'mediaContent') | null;
+                  mediaSize?: ('oneThird' | 'half' | 'twoThirds') | null;
+                  richText?:
+                    | {
+                        [k: string]: unknown;
+                      }[]
+                    | null;
+                  enableLink?: boolean | null;
+                  link?: {
+                    type?: ('reference' | 'custom') | null;
+                    newTab?: boolean | null;
+                    reference?:
+                      | ({
+                          relationTo: 'pages';
+                          value: string | Page;
+                        } | null)
+                      | ({
+                          relationTo: 'projects';
+                          value: string | Project;
+                        } | null);
+                    url?: string | null;
+                    label: string;
+                    appearance?: ('primary' | 'secondary') | null;
+                  };
+                  media: string | Media;
+                  mediaFit?: ('cover' | 'contain') | null;
+                  id?: string | null;
+                }[]
+              | null;
+            id?: string | null;
+            blockName?: string | null;
+            blockType: 'mediaContent';
+          }
+        | {
+            id?: string | null;
+            blockName?: string | null;
+            blockType: 'profile-cta';
+          }
+        | {
+            project: (string | Project)[];
+            id?: string | null;
+            blockName?: string | null;
+            blockType: 'projectGrid';
+          }
+      )[]
+    | null;
+  slug?: string | null;
   meta?: {
-    title?: string;
-    description?: string;
-    image?: string | Media;
+    title?: string | null;
+    description?: string | null;
+    image?: string | Media | null;
   };
   updatedAt: string;
   createdAt: string;
-  _status?: 'draft' | 'published';
+  _status?: ('draft' | 'published') | null;
 }
+/**
+ * This interface was referenced by `Config`'s JSON-Schema
+ * via the `definition` "projects".
+ */
 export interface Project {
   id: string;
   title: string;
   description: {
     [k: string]: unknown;
   }[];
-  technologiesUsed: string[] | Technology[];
-  featuredImage?: string | Media;
-  layout?: (
-    | {
-        contentFields?: {
-          size?: 'oneThird' | 'half' | 'twoThirds' | 'full';
-          richText?: {
-            [k: string]: unknown;
-          }[];
-          enableLink?: boolean;
-          link?: {
-            type?: 'reference' | 'custom';
-            newTab?: boolean;
-            reference:
+  technologiesUsed: (string | Technology)[];
+  featuredImage?: string | Media | null;
+  layout?:
+    | (
+        | {
+            contentFields?:
               | {
-                  value: string | Page;
-                  relationTo: 'pages';
-                }
+                  size?: ('oneThird' | 'half' | 'twoThirds' | 'full') | null;
+                  richText?:
+                    | {
+                        [k: string]: unknown;
+                      }[]
+                    | null;
+                  enableLink?: boolean | null;
+                  link?: {
+                    type?: ('reference' | 'custom') | null;
+                    newTab?: boolean | null;
+                    reference?:
+                      | ({
+                          relationTo: 'pages';
+                          value: string | Page;
+                        } | null)
+                      | ({
+                          relationTo: 'projects';
+                          value: string | Project;
+                        } | null);
+                    url?: string | null;
+                    label: string;
+                    appearance?: ('primary' | 'secondary') | null;
+                  };
+                  id?: string | null;
+                }[]
+              | null;
+            id?: string | null;
+            blockName?: string | null;
+            blockType: 'content';
+          }
+        | {
+            richText?:
               | {
-                  value: string | Project;
-                  relationTo: 'projects';
-                };
-            url: string;
-            label: string;
-            appearance?: 'primary' | 'secondary';
-          };
-          id?: string;
-        }[];
-        id?: string;
-        blockName?: string;
-        blockType: 'content';
-      }
-    | {
-        richText?: {
-          [k: string]: unknown;
-        }[];
-        form: string | Form;
-        id?: string;
-        blockName?: string;
-        blockType: 'form';
-      }
-    | {
-        mediaFields?: {
-          size?: 'oneThird' | 'half' | 'twoThirds' | 'full';
-          media: string | Media;
-          mediaFit?: 'cover' | 'contain';
-          id?: string;
-        }[];
-        id?: string;
-        blockName?: string;
-        blockType: 'mediaBlock';
-      }
-    | {
-        mediaContentFields?: {
-          alignment?: 'contentMedia' | 'mediaContent';
-          mediaSize?: 'oneThird' | 'half' | 'twoThirds';
-          richText?: {
-            [k: string]: unknown;
-          }[];
-          enableLink?: boolean;
-          link?: {
-            type?: 'reference' | 'custom';
-            newTab?: boolean;
-            reference:
+                  [k: string]: unknown;
+                }[]
+              | null;
+            form: string | Form;
+            id?: string | null;
+            blockName?: string | null;
+            blockType: 'form';
+          }
+        | {
+            mediaFields?:
               | {
-                  value: string | Page;
-                  relationTo: 'pages';
-                }
+                  size?: ('oneThird' | 'half' | 'twoThirds' | 'full') | null;
+                  media: string | Media;
+                  mediaFit?: ('cover' | 'contain') | null;
+                  id?: string | null;
+                }[]
+              | null;
+            id?: string | null;
+            blockName?: string | null;
+            blockType: 'mediaBlock';
+          }
+        | {
+            mediaContentFields?:
               | {
-                  value: string | Project;
-                  relationTo: 'projects';
-                };
-            url: string;
-            label: string;
-            appearance?: 'primary' | 'secondary';
-          };
-          media: string | Media;
-          mediaFit?: 'cover' | 'contain';
-          id?: string;
-        }[];
-        id?: string;
-        blockName?: string;
-        blockType: 'mediaContent';
-      }
-  )[];
-  slug?: string;
+                  alignment?: ('contentMedia' | 'mediaContent') | null;
+                  mediaSize?: ('oneThird' | 'half' | 'twoThirds') | null;
+                  richText?:
+                    | {
+                        [k: string]: unknown;
+                      }[]
+                    | null;
+                  enableLink?: boolean | null;
+                  link?: {
+                    type?: ('reference' | 'custom') | null;
+                    newTab?: boolean | null;
+                    reference?:
+                      | ({
+                          relationTo: 'pages';
+                          value: string | Page;
+                        } | null)
+                      | ({
+                          relationTo: 'projects';
+                          value: string | Project;
+                        } | null);
+                    url?: string | null;
+                    label: string;
+                    appearance?: ('primary' | 'secondary') | null;
+                  };
+                  media: string | Media;
+                  mediaFit?: ('cover' | 'contain') | null;
+                  id?: string | null;
+                }[]
+              | null;
+            id?: string | null;
+            blockName?: string | null;
+            blockType: 'mediaContent';
+          }
+      )[]
+    | null;
+  slug?: string | null;
   role: ('uiUxDesigner' | 'frontEndDeveloper' | 'backEndDeveloper')[];
-  startDate?: string;
-  endDate?: string;
+  startDate?: string | null;
+  endDate?: string | null;
   meta?: {
-    title?: string;
-    description?: string;
-    image?: string | Media;
+    title?: string | null;
+    description?: string | null;
+    image?: string | Media | null;
   };
   updatedAt: string;
   createdAt: string;
-  _status?: 'draft' | 'published';
+  _status?: ('draft' | 'published') | null;
 }
+/**
+ * This interface was referenced by `Config`'s JSON-Schema
+ * via the `definition` "technologies".
+ */
 export interface Technology {
   id: string;
   name: string;
   updatedAt: string;
   createdAt: string;
 }
+/**
+ * This interface was referenced by `Config`'s JSON-Schema
+ * via the `definition` "forms".
+ */
 export interface Form {
   id: string;
   title: string;
-  fields?: (
+  fields?:
+    | (
+        | {
+            name: string;
+            label?: string | null;
+            width?: number | null;
+            required?: boolean | null;
+            id?: string | null;
+            blockName?: string | null;
+            blockType: 'email';
+          }
+        | {
+            message?:
+              | {
+                  [k: string]: unknown;
+                }[]
+              | null;
+            id?: string | null;
+            blockName?: string | null;
+            blockType: 'message';
+          }
+        | {
+            name: string;
+            label?: string | null;
+            width?: number | null;
+            defaultValue?: string | null;
+            required?: boolean | null;
+            id?: string | null;
+            blockName?: string | null;
+            blockType: 'text';
+          }
+        | {
+            name: string;
+            label?: string | null;
+            width?: number | null;
+            defaultValue?: string | null;
+            required?: boolean | null;
+            id?: string | null;
+            blockName?: string | null;
+            blockType: 'textarea';
+          }
+      )[]
+    | null;
+  submitButtonLabel?: string | null;
+  confirmationType?: ('message' | 'redirect') | null;
+  confirmationMessage?:
     | {
-        name: string;
-        label?: string;
-        width?: number;
-        defaultValue?: string;
-        required?: boolean;
-        id?: string;
-        blockName?: string;
-        blockType: 'text';
-      }
-    | {
-        name: string;
-        label?: string;
-        width?: number;
-        defaultValue?: string;
-        required?: boolean;
-        id?: string;
-        blockName?: string;
-        blockType: 'textarea';
-      }
-    | {
-        name: string;
-        label?: string;
-        width?: number;
-        required?: boolean;
-        id?: string;
-        blockName?: string;
-        blockType: 'email';
-      }
-    | {
-        message?: {
-          [k: string]: unknown;
-        }[];
-        id?: string;
-        blockName?: string;
-        blockType: 'message';
-      }
-  )[];
-  submitButtonLabel?: string;
-  confirmationType?: 'message' | 'redirect';
-  confirmationMessage: {
-    [k: string]: unknown;
-  }[];
+        [k: string]: unknown;
+      }[]
+    | null;
   redirect?: {
     url: string;
   };
-  emails?: {
-    emailTo?: string;
-    cc?: string;
-    bcc?: string;
-    replyTo?: string;
-    emailFrom?: string;
-    subject: string;
-    message?: {
-      [k: string]: unknown;
-    }[];
-    id?: string;
-  }[];
+  emails?:
+    | {
+        emailTo?: string | null;
+        cc?: string | null;
+        bcc?: string | null;
+        replyTo?: string | null;
+        emailFrom?: string | null;
+        subject: string;
+        message?:
+          | {
+              [k: string]: unknown;
+            }[]
+          | null;
+        id?: string | null;
+      }[]
+    | null;
   updatedAt: string;
   createdAt: string;
 }
+/**
+ * This interface was referenced by `Config`'s JSON-Schema
+ * via the `definition` "users".
+ */
 export interface User {
   id: string;
   updatedAt: string;
   createdAt: string;
   email: string;
-  resetPasswordToken?: string;
-  resetPasswordExpiration?: string;
-  salt?: string;
-  hash?: string;
-  loginAttempts?: number;
-  lockUntil?: string;
-  password?: string;
+  resetPasswordToken?: string | null;
+  resetPasswordExpiration?: string | null;
+  salt?: string | null;
+  hash?: string | null;
+  loginAttempts?: number | null;
+  lockUntil?: string | null;
+  password?: string | null;
 }
+/**
+ * This interface was referenced by `Config`'s JSON-Schema
+ * via the `definition` "form-submissions".
+ */
 export interface FormSubmission {
   id: string;
   form: string | Form;
-  submissionData?: {
-    field: string;
-    value: string;
-    id?: string;
-  }[];
+  submissionData?:
+    | {
+        field: string;
+        value: string;
+        id?: string | null;
+      }[]
+    | null;
   updatedAt: string;
   createdAt: string;
 }
+/**
+ * This interface was referenced by `Config`'s JSON-Schema
+ * via the `definition` "payload-preferences".
+ */
+export interface PayloadPreference {
+  id: string;
+  user: {
+    relationTo: 'users';
+    value: string | User;
+  };
+  key?: string | null;
+  value?:
+    | {
+        [k: string]: unknown;
+      }
+    | unknown[]
+    | string
+    | number
+    | boolean
+    | null;
+  updatedAt: string;
+  createdAt: string;
+}
+/**
+ * This interface was referenced by `Config`'s JSON-Schema
+ * via the `definition` "payload-migrations".
+ */
+export interface PayloadMigration {
+  id: string;
+  name?: string | null;
+  batch?: number | null;
+  updatedAt: string;
+  createdAt: string;
+}
+/**
+ * This interface was referenced by `Config`'s JSON-Schema
+ * via the `definition` "header".
+ */
 export interface Header {
   id: string;
-  navItems?: {
-    link: {
-      type?: 'reference' | 'custom';
-      newTab?: boolean;
-      reference:
-        | {
-            value: string | Page;
-            relationTo: 'pages';
-          }
-        | {
-            value: string | Project;
-            relationTo: 'projects';
-          };
-      url: string;
-      label: string;
-    };
-    id?: string;
-  }[];
-  updatedAt?: string;
-  createdAt?: string;
+  navItems?:
+    | {
+        link: {
+          type?: ('reference' | 'custom') | null;
+          newTab?: boolean | null;
+          reference?:
+            | ({
+                relationTo: 'pages';
+                value: string | Page;
+              } | null)
+            | ({
+                relationTo: 'projects';
+                value: string | Project;
+              } | null);
+          url?: string | null;
+          label: string;
+        };
+        id?: string | null;
+      }[]
+    | null;
+  updatedAt?: string | null;
+  createdAt?: string | null;
 }
+/**
+ * This interface was referenced by `Config`'s JSON-Schema
+ * via the `definition` "profile".
+ */
 export interface Profile {
   id: string;
   name: string;
-  location?: string;
-  title?: string;
-  aboutMe?: {
-    [k: string]: unknown;
-  }[];
-  profileImage?: string | Media;
+  location?: string | null;
+  title?: string | null;
+  aboutMe?:
+    | {
+        [k: string]: unknown;
+      }[]
+    | null;
+  profileImage?: string | Media | null;
   socialLinks?: {
-    github?: string;
-    linkedin?: string;
-    email?: string;
-    twitter?: string;
+    github?: string | null;
+    linkedin?: string | null;
+    email?: string | null;
+    twitter?: string | null;
   };
-  updatedAt?: string;
-  createdAt?: string;
+  updatedAt?: string | null;
+  createdAt?: string | null;
+}
+
+
+declare module 'payload' {
+  export interface GeneratedTypes extends Config {}
 }
