+++ src/utilities/serverConfig.ts
@@ -1,0 +1,3 @@
+
+
+export const serverUrl = process.env.NEXT_PUBLIC_VERCEL_PROJECT_PRODUCTION_URL ?? 'http://localhost:3000'
