+++ src/utilities/tagRevalidator.ts
@@ -1,6 +1,6 @@
 import type { TypeWithID } from 'payload'
 
-import { serverUrl } from '@/utilities/api'
+import { serverUrl } from '@/utilities/serverConfig'
 import { AfterChangeHook } from 'node_modules/payload/dist/globals/config/types'
 
 export const formatAppURL = ({ doc }: any): string => {
